feat(common-scripts): supports inject from registered locks in dao.withdraw
