+++ .changeset/cold-years-dream.md
@@ -1,0 +1,5 @@
+---
+"@ckb-lumos/common-scripts": minor
+---
+
+feat: support registered lock in dao operations
